+++ es5/lists/parse.js
@@ -84,8 +84,8 @@ export default function parse(q, escapeChar) {
           i++; // if there is a ' opened then ignore the ) until the next '
 
           var charAtI = queryString.charAt(i);
-          if (charAtI == "\\") ignoreNextChar = true; // when we have a backslash \then ignore the next char
-          else if (!ignoreNextChar && (charAtI == "'" || charAtI == '"')) openedApos = !openedApos;else if (!ignoreNextChar && charAtI == ")" && !openedApos) parenthesis.open--;else ignoreNextChar = false;
+          if (charAtI == "\\") ignoreNextChar = true; // when we have a backslash \ then ignore the next char
+          else if (!ignoreNextChar && (charAtI == "'" || charAtI == '"')) openedApos = !openedApos;else if (!ignoreNextChar && charAtI == "(" && !openedApos) parenthesis.open++;else if (!ignoreNextChar && charAtI == ")" && !openedApos) parenthesis.open--;else ignoreNextChar = false;
         }
 
         var lastIndex = factory.length - 1; // concat with the first index
@@ -95,7 +95,9 @@ export default function parse(q, escapeChar) {
           factory[0] += parse(queryString.substring(start + 1, i));
           if (closeOperator != "") factory[0] += "</" + closeOperator + ">";
           closeOperator = "";
-        } else factory[0] = parse(queryString.substring(start + 1, i));
+        } else {
+          factory[0] = parse(queryString.substring(start + 1, i));
+        }
 
         break;