+++ src/main.js
@@ -33,6 +33,7 @@ export default class SharepointPlus {
     this.url=void 0;
     this.module_sprequest=null;
     this.credentialOptions=null;
+    this.authMethod={cookie:null};
     this.proxyweb=null;
   }
 
@@ -44,7 +45,7 @@ export default class SharepointPlus {
 
     @return {String} The current SharepointPlus version
   */
-  getVersion () { return "6.0.4" }
+  getVersion () { return "6.1.0" }
 
   /**
     @name $SP().auth
@@ -53,6 +54,7 @@ export default class SharepointPlus {
     @description Permits to use credentials when doing requests (for Node module only)
 
     @param {Object} credentialOptions Options from https://github.com/s-KaiNet/node-sp-auth
+        @params {String} [method] If we cannot use an authentication from 'node-sp-auth', then we need to define the method here (only 'cookie' is supported for now)
     @return {Object} the current SharepointPlus object
 
     @example
@@ -64,9 +66,20 @@ export default class SharepointPlus {
     var sp = $SP().auth(user1);
     sp.list("My List", "https://web.si.te").get({...});
     sp.list("Other List"; "http://my.sharpoi.nt/other.directory/").update(...);
+
+    // let's say we want to use our own function to set the FedAuth cookie
+    var sp = $SP().auth({method:'cookie', function() {
+      // we need to return the content of our cookie
+      // e.g. if it's a FedAuth we'll return 'FedAuth=YAHeZNEZdfnZEfzfzeKnfze…';
+      return myFunctionToGetFedAuth();
+    });
   */
-  auth (credentialOptions) {
-    this.credentialOptions = credentialOptions;
+  auth (credentialOptions, fct) {
+    if (credentialOptions.method === 'cookie' && typeof fct === "function") {
+      this.authMethod.cookie = fct;
+    } else {
+      this.credentialOptions = credentialOptions;
+    }
     return this;
   }
 
@@ -116,6 +129,7 @@ export default class SharepointPlus {
             case "getContentTypeInfo":
             case "getContentTypes":
             case "getWorkflowID":
+            case "hasPermission":
             case "history":
             case "moderate":
             case "remove":